+++ src/Containers/Settings/AccountCard.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
-import styled from 'styled-components';
 import { SignOutAlt } from '@styled-icons/fa-solid/SignOutAlt';
 import { UserCircle } from '@styled-icons/fa-solid/UserCircle';
-import { BaseStyles } from '../../Utils';
 import { MainInterface, ResponsiveInterface } from '../../Utils/BaseStyles';
-import { Paragraph, SmallText } from '../../Text';
 import { SettingsCard } from '../SettingsCard';
+import { DisplayItem } from '../DisplayItem';
 import { Button } from '../../Inputs/Button';
 
-interface DisplayItemProps extends MainInterface, ResponsiveInterface {
-    label: string;
-    value: string;
-}
 export interface AccountProps
     extends React.HTMLAttributes<HTMLDivElement>,
         MainInterface,
@@ -55,25 +49,3 @@ export const AccountCard: React.FC<AccountProps> = ({
         </SettingsCard>
     );
 };
-const DisplayItem: React.FC<DisplayItemProps> = ({
-    label,
-    value,
-    ...props
-}): React.ReactElement => (
-    <Item {...props}>
-        <SmallText lineHeight="1" size="0.9rem" bold>
-            {label}
-        </SmallText>
-        <Text color="grey" bold>
-            {value}
-        </Text>
-    </Item>
-);
-
-const Item = styled.div`
-    ${(props): string => BaseStyles.Main({ margin: '2px 0', ...props })}
-`;
-
-const Text = styled(Paragraph)`
-    overflow-wrap: break-word;
-`;
