+++ src/Containers/index.ts
@@ -26,3 +26,4 @@ export * from './RankingTable';
 export * from './KitchenCard';
 export * from './HeaderRow';
 export * from './Status';
+export * from './DisplayItem';
